# static/openapi.yaml
openapi: "3.0.3"
info:
  title: AiPayGen API
  description: |
    AI-powered API platform with 244+ tools. Pay-per-call with API keys (Bearer apk_xxx),
    x402 V2 USDC micropayments (multi-chain: Base, Solana, Stellar), or use the free tier (1 call/day per IP).
    Supports autonomous agent micropayments — no API key or signup required.

    Base URL: https://api.aipaygen.com
  version: "1.8.1"
  contact:
    email: hello@aipaygen.com
    url: https://aipaygen.com
  license:
    name: Proprietary
servers:
  - url: https://api.aipaygen.com
    description: Production
security:
  - BearerApiKey: []
  - AgentJWT: []
  - {}
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: "API key with apk_ prefix. Obtain via POST /auth/generate-key or POST /credits/buy."
    AgentJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Agent JWT obtained via /agents/challenge + /agents/verify wallet auth flow."
  schemas:
...
    AiToolResponse:
      type: object
      properties:
        result:
          type: string
        model:
          type: string
        _meta:
          type: object
    SentimentResponse:
      type: object
      properties:
        sentiment:
          type: string
          enum: [positive, negative, neutral, mixed]
        confidence:
          type: number
        model:
          type: string
    ResearchResponse:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        sources:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              url:
                type: string
        model:
          type: string
...
paths:
...
  /credits/buy:
    get:
      tags: [Auth]
      summary: Get available purchase options
      security: []
      responses:
        "200":
          description: Purchase options (Stripe and x402)
    post:
      tags: [Auth]
      summary: Buy API credits
      description: |
        Purchase credits via x402 USDC payment or Stripe checkout.
        Returns a Stripe checkout URL or generates a key if payment is provided.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount_usd:
                  type: number
                  example: 5.0
                label:
                  type: string
                  default: credit-pack
      responses:
        "200":
          description: Checkout URL or new key
        "402":
          description: Payment required

  # ── AI Tools ──────────────────────────────────────────────────────────
...
  /research:
    post:
      tags: [AI Tools]
      summary: Research a topic with web sources
      description: Searches the web, scrapes top results, and synthesizes an answer with citations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  example: "What is x402 protocol?"
      responses:
        "200":
          description: Research answer with sources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResponse"
        "400":
          description: Missing question
        "422":
          description: Could not retrieve sources
  /code:
    post:
      tags: [AI Tools]
      summary: Generate code
      description: Generate code in any language from a natural language description.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: "A function that sorts a list of dictionaries by key"
                language:
                  type: string
                  default: Python
                model:
                  type: string
                  description: "Model to use (e.g. claude-haiku, gpt-4o, deepseek)"
      responses:
        "200":
          description: Generated code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiToolResponse"
        "400":
          description: Missing description
  /summarize:
    post:
      tags: [AI Tools]
      summary: Summarize text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                length:
                  type: string
                  default: bullets
                  description: "Format: bullets, paragraph, one-line"
                model:
                  type: string
      responses:
        "200":
          description: Summary result
        "400":
          description: Missing text
  /translate:
    post:
      tags: [AI Tools]
      summary: Translate text to another language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                language:
                  type: string
                  default: Spanish
                model:
                  type: string
      responses:
        "200":
          description: Translated text
        "400":
          description: Missing text
  /sentiment:
    post:
      tags: [AI Tools]
      summary: Analyze sentiment of text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: Sentiment analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentimentResponse"
        "400":
          description: Missing text
  /extract:
    post:
      tags: [AI Tools]
      summary: Extract structured data from text or URL
      description: Extract fields from unstructured text or scrape a URL and extract data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: "Raw text to extract from"
                url:
                  type: string
                  description: "URL to scrape and extract from"
                schema:
                  type: string
                  description: "Description of what to extract"
                fields:
                  type: array
                  items:
                    type: string
                  description: "List of field names to extract"
                model:
                  type: string
      responses:
        "200":
          description: Extracted data
        "400":
          description: Missing text/url or schema/fields
  /classify:
    post:
      tags: [AI Tools]
      summary: Classify text into categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text, categories]
              properties:
                text:
                  type: string
                categories:
                  type: array
                  items:
                    type: string
                  example: ["positive", "negative", "neutral"]
                model:
                  type: string
      responses:
        "200":
          description: Classification result
        "400":
          description: Missing text or categories
  /vision:
    post:
      tags: [AI Tools]
      summary: Analyze an image with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
                  description: "Image URL (also accepts 'image_url')"
                question:
                  type: string
                  default: "Describe this image in detail"
                model:
                  type: string
      responses:
        "200":
          description: Image analysis result
        "400":
          description: Missing URL
  /rag:
    post:
      tags: [AI Tools]
      summary: Retrieval-augmented generation
      description: Answer a question using provided documents as context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [documents, query]
              properties:
                documents:
                  type: string
                  description: "Document text to use as context"
                query:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: RAG answer with citations
        "400":
          description: Missing documents or query
  /web/search:
    get:
      tags: [AI Tools]
      summary: Web search
      description: Search the web via DuckDuckGo and return structured results.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
        - name: "n"
          in: query
          schema:
            type: integer
            default: 10
            maximum: 25
          description: Number of results
      responses:
        "200":
          description: Search results
        "400":
          description: Missing query
    post:
      tags: [AI Tools]
      summary: Web search (POST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  description: "Also accepts 'q'"
                "n":
                  type: integer
                  default: 10
                  maximum: 25
      responses:
        "200":
          description: Search results
        "400":
          description: Missing query

  # ── Agents ────────────────────────────────────────────────────────────
...
  /memory/set:
    post:
      tags: [Memory]
      summary: Store a memory key-value pair
      description: Requires agent JWT authentication.
      security:
        - AgentJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                value: {}
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Memory stored
        "401":
          description: JWT required
  /memory/get:
    post:
      tags: [Memory]
      summary: Retrieve a memory value by key
      security:
        - AgentJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key:
                  type: string
      responses:
        "200":
          description: Memory value
        "401":
          description: JWT required
        "404":
          description: Key not found
  /memory/search:
    post:
      tags: [Memory]
      summary: Search memory by query string
      security:
        - AgentJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
      responses:
        "200":
          description: Matching memory entries
        "401":
          description: JWT required
  /memory/list:
    post:
      tags: [Memory]
      summary: List all memory keys for the authenticated agent
      security:
        - AgentJWT: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        "200":
          description: List of keys
        "401":
          description: JWT required

  # ── Marketplace ───────────────────────────────────────────────────────
...
  /marketplace/call:
    post:
      tags: [Marketplace]
      summary: Call an agent service from the marketplace
      description: Proxies request to the seller agent. Requires API key.
      security:
        - BearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listing_id]
              properties:
                listing_id:
                  type: string
                payload:
                  type: object
                  description: "Payload to forward to the agent"
      responses:
        "200":
          description: Proxied response from seller agent
        "400":
          description: Missing listing_id
        "404":
          description: Listing not found

  # ── Seller ────────────────────────────────────────────────────────────
...
  /session/start:
    post:
      tags: [Sessions]
      summary: Start a new session
      security:
        - BearerApiKey: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id:
                  type: string
                  default: anonymous
                context:
                  type: object
                ttl_hours:
                  type: integer
                  default: 24
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
...
  /write:
    post:
      tags: [AI Tools]
      summary: Generate written content
      description: Generate blog posts, emails, copy, social media, and more.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                tone:
                  type: string
                  default: professional
                format:
                  type: string
                  description: "blog, email, social, copy"
                model:
                  type: string
      responses:
        "200":
          description: Generated content
        "400":
          description: Missing prompt
  /analyze:
    post:
      tags: [AI Tools]
      summary: Analyze text with a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                question:
                  type: string
                  default: general analysis
                model:
                  type: string
      responses:
        "200":
          description: Analysis result
        "400":
          description: Missing content
  /compare:
    post:
      tags: [AI Tools]
      summary: Compare two texts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text_a, text_b]
              properties:
                text_a:
                  type: string
                text_b:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: Comparison with similarities, differences, scores
...